Extract ExecutionTabContainer so tab-specific hooks only run when needed
