+++ src/modules/executions/feature/ExecutionContainer.tsx
@@ -1,37 +1,25 @@
-import { checkpointsQueryKeys } from "@/modules/checkpoints/business-logic/checkpoints-queries";
 import {
 	getCheckpointsPollingInterval,
 	useCheckpoints,
 } from "@/modules/checkpoints/business-logic/use-checkpoints";
-import { useTimelineEntries } from "../business-logic/use-timeline-entries";
 import { CheckpointDetailPanelContainer } from "@/modules/checkpoints/feature/CheckpointDetailPanelContainer";
 import type { PanelTab } from "@/modules/checkpoints/ui/CheckpointDetailPanelTabs";
 import { useManualRefresh } from "@/shared/business-logic/use-manual-refresh";
-import { CopyCommand } from "@/shared/ui/CopyCommand";
 import { RefreshButton } from "@/shared/ui/RefreshButton";
-import { StatusDot } from "@/shared/ui/StatusDot";
 import { ThreePanelLayout } from "@/shared/ui/ThreePanelLayout";
-import {
-	ThreePanelLayoutProvider,
-	useThreePanelLayout,
-} from "@/shared/ui/ThreePanelLayoutContext";
-import { useQueryClient } from "@tanstack/react-query";
+import { ThreePanelLayoutProvider } from "@/shared/ui/ThreePanelLayoutContext";
 import { useNavigate, useParams, useSearch } from "@tanstack/react-router";
 import { useState } from "react";
-import { toast } from "sonner";
-import { executionsQueryKeys } from "../business-logic/executions-queries";
 import { useExecution } from "../business-logic/use-execution";
 import { useExecutions } from "../business-logic/use-executions";
-import { useResolveWaitCondition } from "../business-logic/use-resolve-wait-condition";
 import { useSyncExecutionStatus } from "../business-logic/use-sync-execution-status";
-import { useWaitCondition } from "../business-logic/use-wait-condition";
 import { DEFAULT_EXECUTIONS_POLLING_INTERVAL } from "../domain/fetch-executions";
 import { ExecutionActionsDropdown } from "../ui/ExecutionActionsDropdown";
-import { ExecutionDetails } from "../ui/ExecutionDetails";
 import type { ExecutionLogsScope } from "./ExecutionLogsScopeSidebarContainer";
 import { ExecutionsList } from "../ui/ExecutionsList";
 import { ExecutionTabs, type ExecutionTab } from "../ui/ExecutionTabs";
 import { ExecutionLogsTabContainer } from "./ExecutionLogsTabContainer";
+import { ExecutionTabContainer } from "./ExecutionTabContainer";
 
 const ROUTE_ID = "/_private/_navbar/flows/$flowId/executions/$executionId";
 const ROUTE_PATH = "/flows/$flowId/executions/$executionId";
@@ -84,43 +72,11 @@ function ExecutionContainerBody() {
 			refetchInterval: getCheckpointsPollingInterval,
 		}
 	);
-	const { waitConditionData } = useWaitCondition(
-		executionData?.activeWaitConditionEntry?.id
-	);
-
-	const { timelineEntries } = useTimelineEntries(
-		executionId,
-		checkpointsData.checkpoints
-	);
 
 	useSyncExecutionStatus(
 		checkpointsData.executionStatus,
 		checkpointsData.hasPendingWaitConditionNode
 	);
-
-	const queryClient = useQueryClient();
-	function invalidateExecutionQueries() {
-		queryClient.invalidateQueries({
-			queryKey: executionsQueryKeys.all(flowId),
-		});
-		queryClient.invalidateQueries({
-			queryKey: executionsQueryKeys.detail(executionId),
-		});
-		queryClient.invalidateQueries({
-			queryKey: executionsQueryKeys.waitConditions(executionId),
-		});
-		queryClient.invalidateQueries({
-			queryKey: checkpointsQueryKeys.all(executionId),
-		});
-	}
-
-	const { resolveWaitCondition } = useResolveWaitCondition({
-		onSuccess: invalidateExecutionQueries,
-		onError: () => {
-			invalidateExecutionQueries();
-			toast.error("Failed to resolve wait condition");
-		},
-	});
 
 	const { refresh: refreshExecutionData, isPending: isManualRefreshPending } =
 		useManualRefresh(async () => {
@@ -136,58 +92,10 @@ function ExecutionContainerBody() {
 	>();
 	const [activeCheckpointTab, setActiveCheckpointTab] =
 		useState<PanelTab>("logs");
-	const { expandRight } = useThreePanelLayout();
 
 	const executionsSortedByCreatedAtDesc = [...executionsData].sort((a, b) => {
 		return (b.createdAt?.getTime() ?? 0) - (a.createdAt?.getTime() ?? 0);
 	});
-
-	const shouldShowResumeHint =
-		executionData?.status === "paused" &&
-		!executionData?.activeWaitConditionEntry;
-
-	const resumeHint = shouldShowResumeHint ? (
-		<div className="bg-card flex flex-col">
-			<div className="flex shrink-0 flex-col gap-4 px-4 py-4">
-				<div className="flex items-center gap-2">
-					<StatusDot status="paused" />
-					<span className="text-foreground truncate font-mono text-xs font-semibold">
-						Execution paused
-					</span>
-				</div>
-				<div className="flex flex-col gap-1">
-					<span className="text-muted-foreground text-xs">
-						Resume by running this command in your Kitaru CLI:
-					</span>
-					<CopyCommand
-						code={`kitaru executions resume --exec-id ${executionId}`}
-					/>
-				</div>
-			</div>
-		</div>
-	) : null;
-
-	const centerBody = isLogsTab ? (
-		<ExecutionLogsTabContainer
-			execution={executionData}
-			checkpoints={checkpointsData.checkpoints}
-			selectedScope={selectedScope}
-			onSelectScope={setSelectedScope}
-			onBack={() => setActiveTab("execution")}
-		/>
-	) : (
-		<ExecutionDetails
-			key={executionId}
-			timelineEntries={timelineEntries}
-			onSelectCheckpoint={(id) => {
-				setSelectedCheckpointId(id);
-				expandRight();
-			}}
-			waitCondition={waitConditionData}
-			onResolveWaitCondition={resolveWaitCondition}
-			resumeHint={resumeHint}
-		/>
-	);
 
 	return (
 		<div className="flex flex-1 flex-col overflow-hidden">
@@ -211,7 +119,25 @@ function ExecutionContainerBody() {
 						activeexecutionId={executionId}
 					/>
 				}
-				center={centerBody}
+				center={
+					isLogsTab ? (
+						<ExecutionLogsTabContainer
+							execution={executionData}
+							checkpoints={checkpointsData.checkpoints}
+							selectedScope={selectedScope}
+							onSelectScope={setSelectedScope}
+							onBack={() => setActiveTab("execution")}
+						/>
+					) : (
+						<ExecutionTabContainer
+							executionId={executionId}
+							flowId={flowId}
+							execution={executionData}
+							checkpoints={checkpointsData.checkpoints}
+							onSelectCheckpoint={setSelectedCheckpointId}
+						/>
+					)
+				}
 				right={
 					isLogsTab ? null : (
 						<CheckpointDetailPanelContainer
